Made lrSlopeMultiplier float

+++ Sources/Models/LangMotion/OptimizerWrapper.swift
@@ -10,11 +10,11 @@ public struct OptimizerOpts {
     public let beta2: Float
     public let weightDecayRate: Float
     public let useBiasCorrection: Bool
-    public let lrSlopeMultiplier: Int
+    public let lrSlopeMultiplier: Float
     public let nEpochs: Int
     public var stepsPerEpoch: Int
     
-    public init(peakLearningRate: Float = 5e-4, beta1: Float = 0.9, beta2: Float = 0.999, weightDecayRate: Float, useBiasCorrection: Bool = false, lrSlopeMultiplier: Int = 1, nEpochs: Int = 10, stepsPerEpoch: Int = 1) {
+    public init(peakLearningRate: Float = 5e-4, beta1: Float = 0.9, beta2: Float = 0.999, weightDecayRate: Float, useBiasCorrection: Bool = false, lrSlopeMultiplier: Float = 1.0, nEpochs: Int = 10, stepsPerEpoch: Int = 1) {
         self.peakLearningRate = peakLearningRate
         self.beta1 = beta1
         self.beta2 = beta2
@@ -50,7 +50,7 @@ public class OptimizerWrapper {
             baseParameter: FixedParameter<Float>(opts.peakLearningRate),
               warmUpStepCount: 20,
               warmUpOffset: 0),
-            slope: -(opts.peakLearningRate / Float(opts.stepsPerEpoch * opts.nEpochs * opts.lrSlopeMultiplier)),  // The LR decays linearly to zero.
+            slope: -(opts.peakLearningRate / Float(Float(opts.stepsPerEpoch) * Float(opts.nEpochs) * opts.lrSlopeMultiplier)),  // The LR decays linearly to zero.
           startStep: 10
         )
     }
